test_automation/requirements.txt
Appium-Python-Client==1.2.0
selenium==3.141.0
urllib3==1.26.6
behave
ipdb
page_objects
firebase-admin